rename unit test class and update some method test logic
