+++ library/src/test/java/com/midtrans/java/MidtransIrisApiTest.java
@@ -1,10 +1,10 @@
 package com.midtrans.java;
 
 import com.midtrans.Config;
-import com.midtrans.Midtrans;
-import com.midtrans.httpclient.IrisApi;
+import com.midtrans.ConfigFactory;
 import com.midtrans.httpclient.error.MidtransError;
 import com.midtrans.java.mockupdata.DataMockup;
+import com.midtrans.service.MidtransIrisApi;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.junit.Before;
@@ -15,107 +15,100 @@ import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.UUID;
 
 import static com.midtrans.java.mockupdata.Constant.approverKey;
 import static com.midtrans.java.mockupdata.Constant.creatorKey;
 import static com.midtrans.java.mockupdata.DataMockup.refNumber;
 
 
-public class IrisApiStaticTest {
-    private DataMockup dataMockup;
-    private Map<String, String> beneficiaries = new HashMap<>();
-    private Config configOptions;
+public class MidtransIrisApiTest {
+    MidtransIrisApi irisApi;
+    DataMockup dataMockup;
+    Map<String, String> beneficiaries = new HashMap<>();
 
 
     @Before
     public void setUp() {
         dataMockup = new DataMockup();
-
-        configOptions = Config.builder()
-                .setSERVER_KEY(creatorKey)
-                .enableLog(false)
-                .build();
+        irisApi = new ConfigFactory(new Config(creatorKey, null, false)).getIrisApi();
     }
 
     @Test
     public void ping() throws MidtransError {
-        String result = IrisApi.ping();
+        String result = irisApi.ping();
         assert result.equals("pong");
     }
 
     @Test
     public void getAggregatorBalance() throws MidtransError {
-        JSONObject result = IrisApi.getBalance(configOptions);
+        JSONObject result = irisApi.getBalance();
         assert result.has("balance");
     }
 
     @Test
     public void createUpdateGetBeneficiaries() throws MidtransError {
         beneficiaries = dataMockup.initDataBeneficiaries();
-
-        /* Using method with Global Config */
-        JSONObject result = IrisApi.createBeneficiaries(beneficiaries, configOptions);
+        JSONObject result = irisApi.createBeneficiaries(beneficiaries);
         assert result.has("status");
         assert result.getString("status").equals("created");
-        getListAndUpdateBeneficiaries(configOptions);
 
+        getListBeneficiaries();
+        updateBeneficiaries();
     }
 
-    private void getListAndUpdateBeneficiaries(Config config) throws MidtransError {
-        JSONArray result2 = IrisApi.getBeneficiaries(config);
-        JSONObject jsonResult2 = new JSONObject(result2.get(result2.length() - 1).toString());
-        assert jsonResult2.getString("alias_name").equals(beneficiaries.get("alias_name"));
+    private void updateBeneficiaries() throws MidtransError {
         String oldAliasName = beneficiaries.get("alias_name");
-        updateBeneficiaries(oldAliasName, config);
+        beneficiaries.replace("alias_name", oldAliasName, oldAliasName + "edt");
+        JSONObject result = irisApi.updateBeneficiaries(oldAliasName, beneficiaries);
+        assert result.getString("status").equals("updated");
     }
 
-    private void updateBeneficiaries(String oldAliasName, Config config) throws MidtransError {
-        beneficiaries.replace("alias_name", oldAliasName, oldAliasName + "edt");
-        JSONObject result = IrisApi.updateBeneficiaries(oldAliasName, beneficiaries, config);
-        assert result.getString("status").equals("updated");
+    private void getListBeneficiaries() throws MidtransError {
+        JSONArray result = irisApi.getBeneficiaries();
+        JSONObject jsonResult = new JSONObject(result.get(result.length() - 1).toString());
+        assert jsonResult.getString("alias_name").equals(beneficiaries.get("alias_name"));
     }
 
 
     @Test
     public void createAndApprovePayouts() throws MidtransError {
-        JSONObject beneficiaries = getRandomBeneficiaries(configOptions);
-
-
-        JSONObject response = IrisApi.createPayouts(dataMockup.initDataRequestPayout(beneficiaries), configOptions);
-        assert response.has("payouts");
+        JSONObject beneficiaries = getRandomBeneficiaries();
 
         /*
-        Approve payouts to Iris API
+        Request payout to Iris API
          */
-        Map<String, Object> approvePayout = new HashMap<>();
-        approvePayout.put("reference_nos", getPayoutsFromResult(response));
-        approvePayout.put("otp", "335163");
+        JSONObject response = irisApi.createPayouts(dataMockup.initDataRequestPayout(beneficiaries));
+        assert response.has("payouts");
 
-        configOptions.setSERVER_KEY(approverKey);
-        JSONObject result = IrisApi.approvePayouts(approvePayout, configOptions);
-        assert result.getString("status").equals("ok");
-    }
-
-    private ArrayList<String> getPayoutsFromResult(JSONObject response) {
         JSONArray responsePayouts = response.getJSONArray("payouts");
         responsePayouts.get(0);
 
         ArrayList<String> referenceNos = new ArrayList<>();
         referenceNos.add(responsePayouts.getJSONObject(0).getString("reference_no"));
-        return referenceNos;
+
+        /*
+        Approve payouts to Iris API
+         */
+        Map<String, Object> approvePayout = new HashMap<>();
+        approvePayout.put("reference_nos", referenceNos);
+        approvePayout.put("otp", "335163");
+        approvePayouts(approvePayout);
+    }
+
+    private void approvePayouts(Map<String, Object> params) throws MidtransError {
+        irisApi.apiConfig().setSERVER_KEY(approverKey);
+        JSONObject result = irisApi.approvePayouts(params);
+        assert result.getString("status").equals("ok");
     }
 
     @Test
     public void createAndRejectPayouts() throws MidtransError {
-        JSONObject beneficiaries = getRandomBeneficiaries(configOptions);
-
-        Midtrans.serverKey = creatorKey;
+        JSONObject beneficiaries = getRandomBeneficiaries();
 
         /*
         Request payout to Iris API
          */
-        JSONObject response = IrisApi.createPayouts(dataMockup.initDataRequestPayout(beneficiaries), configOptions);
+        JSONObject response = irisApi.createPayouts(dataMockup.initDataRequestPayout(beneficiaries));
         assert response.has("payouts");
 
         JSONArray responsePayouts = response.getJSONArray("payouts");
@@ -134,32 +127,30 @@ public class IrisApiStaticTest {
     }
 
     private void rejectPayouts(Map<String, Object> params) throws MidtransError {
-        configOptions.setSERVER_KEY(approverKey);
-        JSONObject result = IrisApi.rejectPayouts(params, configOptions);
+        irisApi.apiConfig().setSERVER_KEY(approverKey);
+        JSONObject result = irisApi.rejectPayouts(params);
         assert result.getString("status").equals("ok");
     }
 
     @Test
     public void getTransactionHistory() throws MidtransError {
-        Midtrans.serverKey = creatorKey;
         LocalDate localDate = LocalDate.now();
         String fromDate = DateTimeFormatter.ofPattern("yyy-MM-dd").format(localDate.minusMonths(1));
         String toDate = DateTimeFormatter.ofPattern("yyy-MM-dd").format(localDate);
-        JSONArray arrayResult = IrisApi.getTransactionHistory(fromDate, toDate, configOptions);
+        JSONArray arrayResult = irisApi.getTransactionHistory(fromDate, toDate);
         assert !arrayResult.isEmpty();
     }
 
     @Test
     public void getTopUpChannels() throws MidtransError {
-        Midtrans.serverKey = creatorKey;
-        JSONArray result = IrisApi.getTopUpChannels(configOptions);
+        JSONArray result = irisApi.getTopUpChannels();
         JSONObject jsonResult = new JSONObject(result.get(0).toString());
         assert jsonResult.getString("virtual_account_type").equals("mandiri_bill_key");
     }
 
     @Test
     public void getBeneficiaryBanks() throws MidtransError {
-        JSONObject result = IrisApi.getBeneficiaryBanks(configOptions);
+        JSONObject result = irisApi.getBeneficiaryBanks();
         JSONArray beneficiaryBanks = result.getJSONArray("beneficiary_banks");
         JSONObject arrayList = null;
         for (int i = 0; i < beneficiaryBanks.length(); i++) {
@@ -174,7 +165,7 @@ public class IrisApiStaticTest {
 
     @Test
     public void validateBankAccount() throws MidtransError {
-        JSONObject result = IrisApi.validateBankAccount("danamon", "000001137298", configOptions);
+        JSONObject result = irisApi.validateBankAccount("danamon", "000001137298");
         assert result.getString("account_no").equals("000001137298");
     }
 
@@ -182,42 +173,38 @@ public class IrisApiStaticTest {
     public void createBeneficiaryWithIdempotencyKey() throws MidtransError {
         Map beneficiary = dataMockup.initDataBeneficiaries();
 
+        // Use idempotency key
+        irisApi.apiConfig().setIrisIdempotencyKey("Iris-Idempotency-Key" + dataMockup.random());
+
         // Request create beneficiary to Iris API
-        Config configOptions = Config.builder()
-                .setSERVER_KEY(creatorKey)
-                .setIrisIdempotencyKey(String.valueOf(UUID.randomUUID()))
-                .build();
-        JSONObject result1 = IrisApi.createBeneficiaries(beneficiary, configOptions);
+        JSONObject result1 = irisApi.createBeneficiaries(beneficiary);
         assert result1.getString("status").equals("created");
 
-        JSONObject result2 = IrisApi.createBeneficiaries(beneficiary, configOptions);
+        JSONObject result2 = irisApi.createBeneficiaries(beneficiary);
         assert result2.getString("status").equals("created");
     }
 
     @Test
-    public void failureCredentials() {
-        configOptions.setSERVER_KEY("dummy");
-        try {
-            IrisApi.getBalance(configOptions);
-        } catch (MidtransError e) {
-            e.printStackTrace();
-            assert e.getResponseBody().equals("HTTP Basic: Access denied.\n");
-        }
+    public void failureCredentials() throws MidtransError {
+        irisApi.apiConfig().setSERVER_KEY(".");
+        JSONObject result = irisApi.getBalance();
+        assert result.getString("HTTP Basic").equals("Access denied.");
     }
 
     @Test
     public void failureGetPayoutsDetails() {
-        JSONObject result;
+        irisApi.apiConfig().setSERVER_KEY(approverKey);
+        JSONObject result = null;
         try {
-            result = IrisApi.getPayoutDetails(refNumber, configOptions);
-            assert result.getString("error_message").equals("Specified payout not found");
-        } catch (MidtransError MidtransError) {
-            MidtransError.printStackTrace();
+            result = irisApi.getPayoutDetails(refNumber);
+        } catch (MidtransError midtransError) {
+            midtransError.printStackTrace();
         }
+        assert result.getString("error_message").equals("Specified payout not found");
     }
 
-    private JSONObject getRandomBeneficiaries(Config config) throws MidtransError {
-        JSONArray result = IrisApi.getBeneficiaries(config);
+    private JSONObject getRandomBeneficiaries() throws MidtransError {
+        JSONArray result = irisApi.getBeneficiaries();
         return new JSONObject(result.get(result.length() - 2).toString());
     }
